requirements.txt
wheel==0.36.2
django==2.2.10
django-mptt
googletrans==2.4.0
django-autotranslate==1.1.1
django-ckeditor==5.9.0
djangorestframework
markdown
django-filter
Pillow
celery==5.0.5
jsonfield==3.1.0
django-compressor==2.4
psycopg2
textblob
unidecode
beautifulsoup4
docxtpl
num2words
singlemodeladmin
python-Levenshtein==0.12.0
fuzzywuzzy==0.18.0
django-datatables-view==1.19.1
xlsxwriter